Cleaned up thread code.
Separated the thread code from the Floor class and added it to its own
class.
